requirements_production.txt
# Production Requirements - Multi-Tenant EFRIS API

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.3

# Existing Dependencies (QuickBooks will be added later)
requests==2.31.0
cryptography==42.0.0
pycryptodome==3.20.0
